TCP server fix

+++ SpeedTester/SpeedTester/Model/Server/ServerStats.cs
@@ -5,7 +5,7 @@ namespace SpeedTester.Model
    public  class ServerStats
     {
         public int DataSize { get; set; } = 0;
-        public int TotalSize { get; set; } = 0;
+        public long TotalSize { get; set; } = 0;
         public int TransmissionTime { get; set; } = 0;
         public int TransmissionSpeed { get; set; } = 0;
         public int LostData { get; set; } = 0;
